chore: move dependencies into a sub-folder
