+++ src/closure-webpack-plugin.js
@@ -7,17 +7,17 @@ const {
 const { ConcatSource, SourceMapSource } = require('webpack-sources');
 const RequestShortener = require('webpack/lib/RequestShortener');
 const Entrypoint = require('webpack/lib/Entrypoint');
-const HarmonyImportDependencyTemplate = require('./harmony-import-dependency-template');
-const HarmonyImportSpecifierDependencyTemplate = require('./harmony-import-specifier-dependency-template');
-const HarmonyNoopTemplate = require('./harmony-noop-template');
-const ImportDependencyTemplate = require('./import-dependency-template');
-const AMDDefineDependencyTemplate = require('./amd-define-dependency-template');
+const HarmonyImportDependencyTemplate = require('./dependencies/harmony-import-dependency-template');
+const HarmonyImportSpecifierDependencyTemplate = require('./dependencies/harmony-import-specifier-dependency-template');
+const HarmonyNoopTemplate = require('./dependencies/harmony-noop-template');
+const ImportDependencyTemplate = require('./dependencies/import-dependency-template');
+const AMDDefineDependencyTemplate = require('./dependencies/amd-define-dependency-template');
 const GoogRequireParserPlugin = require('./goog-require-parser-plugin');
-const GoogDependency = require('./goog-dependency');
-const GoogLoaderPrefixDependency = require('./goog-loader-prefix-dependency');
-const GoogLoaderSuffixDependency = require('./goog-loader-suffix-dependency');
-const GoogLoaderEs6PrefixDependency = require('./goog-loader-es6-prefix-dependency');
-const GoogLoaderEs6SuffixDependency = require('./goog-loader-es6-suffix-dependency');
+const GoogDependency = require('./dependencies/goog-dependency');
+const GoogLoaderPrefixDependency = require('./dependencies/goog-loader-prefix-dependency');
+const GoogLoaderSuffixDependency = require('./dependencies/goog-loader-suffix-dependency');
+const GoogLoaderEs6PrefixDependency = require('./dependencies/goog-loader-es6-prefix-dependency');
+const GoogLoaderEs6SuffixDependency = require('./dependencies/goog-loader-es6-suffix-dependency');
 const NullFactory = require('webpack/lib/NullFactory');
 const validateOptions = require('schema-utils');
 const closureWebpackPluginSchema = require('../schema/plugin.json');
