+++ src/dependencies/amd-define-dependency-template.js
