+++ src/dependencies/goog-dependency.js
