+++ src/dependencies/goog-loader-es6-prefix-dependency.js
