+++ src/dependencies/harmony-import-dependency-template.js
