+++ src/dependencies/harmony-import-specifier-dependency-template.js
