+++ src/dependencies/harmony-noop-template.js
