+++ src/dependencies/import-dependency-template.js
